judge_server/requirements.txt
asgiref==3.4.1
async-timeout==4.0.2
certifi==2022.6.15
charset-normalizer==2.0.12
Deprecated==1.2.13
Django==3.2.14
django-redis==5.2.0
idna==3.3
importlib-metadata==4.8.3
packaging==21.3
PyMySQL==1.0.2
pyparsing==3.0.9
pytz==2022.1
redis==4.3.4
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
thrift==0.16.0
typing_extensions==4.1.1
urllib3==1.26.9
websocket-client==1.3.1
wrapt==1.14.1
zipp==3.6.0
